Se cambia la logica de redireccion al entrar a la mesa
