+++ Foody/src/app/pages/login/login.page.html
@@ -1,4 +1,4 @@
-<ion-content [fullscreen]="true" *ngIf="!this.logged" style="--background: #FDDCA1;">
+<ion-content [fullscreen]="true" *ngIf="!this.logged">
   <div class="container">
     <div class="blocks up-login" style="flex: 2;">
       <div class="title-container">
